adapting plot to handle 0-values
Initially had issues with merging on a column that didn't exist (see BLAST results commit on build). While testing with a small dataset I then ran into problems with the denominator of the F1 score calculation turning 0. This shouldn't really happen with a dataset of reasonable size for building a model, but that would have taken far too many resources for testing locally. Most notably, I adapted the plotting in such a way that it can handle 0-values, which turned out to be very helpful as it makes gaps in the alignments very obvious. Also, the labels were turned sideways for better legibility.